refactor: update import paths in mocks and tests for consistency
